+++ src/lib/index.ts
@@ -1,2 +1,3 @@
 export * from './utils';
 export * from './constant';
+export * from './calendar-utils';
